Move implementations from impl/*.h to *_impl.h
