+++ Makefile.am
@@ -7,30 +7,30 @@ libsecp256k1_la_LIBADD = $(SECP_LIBS)
 include_HEADERS = include/secp256k1.h
 noinst_HEADERS =
 noinst_HEADERS += src/group.h
+noinst_HEADERS += src/group_impl.h
 noinst_HEADERS += src/num_openssl.h
+noinst_HEADERS += src/num_openssl_impl.h
 noinst_HEADERS += src/num_gmp.h
+noinst_HEADERS += src/num_gmp_impl.h
 noinst_HEADERS += src/ecdsa.h
+noinst_HEADERS += src/ecdsa_impl.h
 noinst_HEADERS += src/ecmult.h
+noinst_HEADERS += src/ecmult_impl.h
 noinst_HEADERS += src/num.h
+noinst_HEADERS += src/num_impl.h
 noinst_HEADERS += src/field_10x26.h
+noinst_HEADERS += src/field_10x26_impl.h
 noinst_HEADERS += src/field_5x52.h
+noinst_HEADERS += src/field_5x52_impl.h
+noinst_HEADERS += src/field_5x52_int128_impl.h
+noinst_HEADERS += src/field_5x52_asm_impl.h
 noinst_HEADERS += src/java/org_bitcoin_NativeSecp256k1.h
 noinst_HEADERS += src/util.h
+noinst_HEADERS += src/util_impl.h
 noinst_HEADERS += src/field_gmp.h
-noinst_HEADERS += src/impl/group.h
-noinst_HEADERS += src/impl/num_openssl.h
-noinst_HEADERS += src/impl/num_gmp.h
-noinst_HEADERS += src/impl/ecdsa.h
-noinst_HEADERS += src/impl/field_5x52_int128.h
-noinst_HEADERS += src/impl/ecmult.h
-noinst_HEADERS += src/impl/num.h
-noinst_HEADERS += src/impl/field_10x26.h
-noinst_HEADERS += src/impl/field_5x52.h
-noinst_HEADERS += src/impl/util.h
-noinst_HEADERS += src/impl/field_gmp.h
-noinst_HEADERS += src/impl/field_5x52_asm.h
-noinst_HEADERS += src/impl/field.h
+noinst_HEADERS += src/field_gmp_impl.h
 noinst_HEADERS += src/field.h
+noinst_HEADERS += src/field_impl.h
 
 if USE_ASM
 libsecp256k1_la_SOURCES += src/field_5x52_asm.asm
