+++ src/bench.c
@@ -4,12 +4,12 @@
 
 #include <stdio.h>
 
-#include "impl/num.h"
-#include "impl/field.h"
-#include "impl/group.h"
-#include "impl/ecmult.h"
-#include "impl/ecdsa.h"
-#include "impl/util.h"
+#include "num_impl.h"
+#include "field_impl.h"
+#include "group_impl.h"
+#include "ecmult_impl.h"
+#include "ecdsa_impl.h"
+#include "util_impl.h"
 
 void random_num_order(secp256k1_num_t *num) {
     do {
