+++ src/ecdsa_impl.h
@@ -5,11 +5,11 @@
 #ifndef _SECP256K1_ECDSA_IMPL_H_
 #define _SECP256K1_ECDSA_IMPL_H_
 
-#include "../num.h"
-#include "../field.h"
-#include "../group.h"
-#include "../ecmult.h"
-#include "../ecdsa.h"
+#include "num.h"
+#include "field.h"
+#include "group.h"
+#include "ecmult.h"
+#include "ecdsa.h"
 
 void static secp256k1_ecdsa_sig_init(secp256k1_ecdsa_sig_t *r) {
     secp256k1_num_init(&r->r);
