+++ src/field_impl.h
@@ -10,11 +10,11 @@
 #endif
 
 #if defined(USE_FIELD_GMP)
-#include "field_gmp.h"
+#include "field_gmp_impl.h"
 #elif defined(USE_FIELD_10X26)
-#include "field_10x26.h"
+#include "field_10x26_impl.h"
 #elif defined(USE_FIELD_5X52)
-#include "field_5x52.h"
+#include "field_5x52_impl.h"
 #else
 #error "Please select field implementation"
 #endif
