+++ src/group_impl.h
@@ -7,9 +7,9 @@
 
 #include <string.h>
 
-#include "../num.h"
-#include "../field.h"
-#include "../group.h"
+#include "num.h"
+#include "field.h"
+#include "group.h"
 
 void static secp256k1_ge_set_infinity(secp256k1_ge_t *r) {
     r->infinity = 1;
