+++ src/num_gmp_impl.h
